requirements.txt
fastapi
azure-functions
mangum
uvicorn
